+++ src/upnp.http.server.cpp
@@ -105,7 +105,7 @@ Server::Server(uv::Loop& loop, const uv::Address& address)
                     {
                         if (nread != UV_EOF)
                         {
-                            log::error("Failed to read from socket {}", uv::getErrorString(nread));
+                            log::error("Failed to read from socket {}", uv::getErrorString(static_cast<int32_t>(nread)));
                         }
                         else
                         {
